Register sink function for floating objects
The Python GObject bindings need to register a sink function for the initially
unowned object classes, so that they can take ownership of the objects and
garbage collect them safely. Clutter has two classes with the initial floating
reference, ClutterActor and ClutterAlpha. This patch registers a sink function
for both of them. Thanks to Lauri Taimila for spotting this issue.