Check for PyInt and PyLong in Actor.animate*
The value for AnimationMode can be PyInt (in case of
clutter.AnimationMode) and PyLong (in case of a custom mode returned by
alpha_register_func).

Also generate a more useful exception for people who try to use the
C-ish _animate(..., "signal::completed", ...) API. There is really no
reason to support this in Python.