Override __str__ and __repr__ for clutter.Path
Since a clutter.Path is easily stringified we should provide an override
for the __str__ and __repr__ slots of clutter.Path. The new code is the
equivalent of:

        def __str__(path):
            return path.get_description()

        def __repr__(path):
            return "clutter.Path(description=\'%s\')" % (str(path))

And should help with debugging paths.

Also, we should split out the ClutterPath overrides into a separate
file, to avoid cluttering the Behaviour overrides.