[script] Add overrides for ClutterScriptable
Allow users to override the do_set_custom_property and
do_parse_custom_node methods. It's now possible to use custom Python
actors in a Clutter Json script.
The signature of do_parse_custom_node is:

  do_parse_custom_node(self, script, name, node)

where 'node' is the JsonNode turned into a native Python object. A
JsonArray is a list, a JsonObject is a dictionary, everything else will
be converted it's Python equivalent. The return value of this method has
to be either a simple Python type (int, float, str) or a gobject.GObject
(or GBoxed, GEnum, ...) derived type. To indicate that you're not able
to handle this node you have to return None. The best way to do this is
to simply chain up to the actors parent implementation and return this
value.